Self-Querying Retrieval/backend/requirements.txt
lark 
langchain-chroma 
langchain_chroma
langchain 
langchain_huggingface 
langchain-openai 
langchain_core
langchain-community 
sentence-transformers
langchain-qdrant
pyodbc
ipykernel
pandas
numpy
chromadb
fastapi
uvicorn
pydantic
aiofiles
attrs>=23.2.0
azure-identity
azure-storage-blob
azure-storage-file-share
google-auth==2.29.0
googleapis-common-protos==1.62.0
python-jose[cryptography]
pyyaml==6.0.1
requests==2.31.0
cffi==1.15.1